# contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "aiwatcher",
    "description": "Observability for agent runs: events in, traces and live streams out.",
    "license": {
      "name": "Apache-2.0",
      "identifier": "Apache-2.0"
    },
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/conversations": {
      "get": {
        "tags": [
          "runs"
        ],
        "summary": "Sessions, each grouping the runs it produced.",
        "description": "The level above a run, so the panel can start from \"which session\" and walk\ndown to a single LLM call without going back to a flat list and re-filtering.",
        "operationId": "list_conversations",
        "parameters": [
          {
            "name": "window_seconds",
            "in": "query",
            "description": "Only sessions with activity in the last this-many seconds. See\n[`crate::window`].",
            "required": false,
            "schema": {
              "type": [
                "integer",
                "null"
              ],
              "format": "int64"
            }
          },
          {
            "name": "agent_id",
            "in": "query",
            "required": false,
            "schema": {
              "type": [
                "string",
                "null"
              ]
            }
          },
          {
            "name": "search",
            "in": "query",
            "description": "Substring match on the conversation id. The one control that turns a\nlong list into the session someone is actually looking for.",
            "required": false,
            "schema": {
              "type": [
                "string",
                "null"
              ]
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": [
                "integer",
                "null"
              ],
              "minimum": 0
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConversationPage"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/dimensions/{kind}": {
      "get": {
        "tags": [
          "runs"
        ],
        "summary": "One dimension's rows: the explorer's top level, whatever it is rooted on.",
        "description": "`session`, `agent`, `runtime`, `workflow`, `trace`, `model` and `tool` all\nreturn the same row shape, so the panel's tree has one renderer rather than\none per pivot. See `aiwatcher_projector::dimensions`.",
        "operationId": "list_dimension",
        "parameters": [
          {
            "name": "kind",
            "in": "path",
            "description": "What to group runs by",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/DimensionKind"
            }
          },
          {
            "name": "window_seconds",
            "in": "query",
            "description": "Only runs with activity in the last this-many seconds, before they are\ngrouped. See [`crate::window`] — a row whose every run falls outside the\nwindow disappears with them rather than staying behind as an empty key.",
            "required": false,
            "schema": {
              "type": [
                "integer",
                "null"
              ],
              "format": "int64"
            }
          },
          {
            "name": "agent_id",
            "in": "query",
            "description": "Narrow to runs that ran this agent, whatever the dimension is. Lets the\ntree stay scoped when someone arrives from an agent-filtered view.",
            "required": false,
            "schema": {
              "type": [
                "string",
                "null"
              ]
            }
          },
          {
            "name": "search",
            "in": "query",
            "description": "Substring match on the key. The one control that turns a long list into\nthe row someone is looking for.",
            "required": false,
            "schema": {
              "type": [
                "string",
                "null"
              ]
            }
          },
          {
            "name": "after",
            "in": "query",
            "description": "Cursor: return rows after this key in the current sort order. Keyed\nrather than offset, because the list reorders as runs arrive.",
            "required": false,
            "schema": {
              "type": [
                "string",
                "null"
              ]
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": [
                "integer",
                "null"
              ],
              "minimum": 0
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DimensionPage"
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/evaluations": {
      "get": {
        "tags": [
          "evaluation"
        ],
        "summary": "Evaluation reports, newest first.",
        "description": "The other half of the loop the traces come from: a trace says what one run\ndid, an evaluation says whether the thing producing those runs is getting\nbetter. They arrive on the same log and are folded apart — an `eval.*`\nevent produces no span and no row in the runs list. See\n`aiwatcher_projector::evaluations`.",
        "operationId": "list_evaluations",
        "parameters": [
          {
            "name": "window_seconds",
            "in": "query",
            "description": "Only reports that finished — or, while they are still running, started\n— in the last this-many seconds. See [`crate::window`].",
            "required": false,
            "schema": {
              "type": [
                "integer",
                "null"
              ],
              "format": "int64"
            }
          },
          {
            "name": "suite",
            "in": "query",
            "required": false,
            "schema": {
              "type": [
                "string",
                "null"
              ]
            }
          },
          {
            "name": "dataset",
            "in": "query",
            "required": false,
            "schema": {
              "type": [
                "string",
                "null"
              ]
            }
          },
          {
            "name": "variant",
            "in": "query",
            "required": false,
            "schema": {
              "type": [
                "string",
                "null"
              ]
            }
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "oneOf": [
                {
                  "type": "null"
                },
                {
                  "$ref": "#/components/schemas/EvaluationStatus"
                }
              ]
            }
          },
          {
            "name": "search",
            "in": "query",
            "description": "Substring over the id, the suite, the dataset, the variant and the\nparameter values.",
            "required": false,
            "schema": {
              "type": [
                "string",
                "null"
              ]
            }
          },
          {
            "name": "after",
            "in": "query",
            "description": "Cursor: the last evaluation id on the previous page. Exclusive.",
            "required": false,
            "schema": {
              "type": [
                "string",
                "null"
              ]
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": [
                "integer",
                "null"
              ],
              "minimum": 0
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EvaluationPage"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/evaluations/{evaluation_id}": {
      "get": {
        "tags": [
          "evaluation"
        ],
        "summary": "One evaluation: its parameters, its metrics, its cases, the document the\nproducer attached, and the previous evaluation of the same suite on the\nsame dataset.",
        "operationId": "get_evaluation",
        "parameters": [
          {
            "name": "evaluation_id",
            "in": "path",
            "description": "The evaluation to fetch",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EvaluationDetail"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/events/stream": {
      "get": {
        "tags": [
          "live"
        ],
        "summary": "Server-sent events for the whole system: a catch-up marker, then live.",
        "description": "This is the panel's Observability transport. It is deliberately separate\nfrom the WebSocket below: the panel only receives here, so EventSource can\nown reconnects and `Last-Event-ID` resume without client-side machinery.",
        "operationId": "stream_events",
        "parameters": [
          {
            "name": "from",
            "in": "query",
            "description": "Resume point. Usually unnecessary for SSE — the browser sends\n`Last-Event-ID` on its own — but explicit here for non-browser clients.",
            "required": false,
            "schema": {
              "type": [
                "string",
                "null"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "text/event-stream of LiveFrame"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ConversationPage": {
        "type": "object",
        "required": [
          "conversations",
          "total",
          "ungrouped_runs"
        ],
        "properties": {
          "conversations": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ConversationSummary"
            }
          },
          "total": {
            "type": "integer",
            "minimum": 0
          },
          "ungrouped_runs": {
            "type": "integer",
            "format": "int64",
            "description": "Runs with no `conversation_id`. They are reachable from the runs list\nbut have no session to sit under, and silently dropping them would make\nthe two views disagree about how much exists.",
            "minimum": 0
          }
        }
      },
...
      "DimensionKind": {
        "type": "string",
        "description": "What the tree is rooted on.",
        "enum": [
          "session",
          "agent",
          "runtime",
          "workflow",
          "trace",
          "model",
          "tool"
        ]
      },
...
      "ErrorBody": {
        "type": "object",
        "description": "The body every error response carries.",
        "required": [
          "code",
          "message"
        ],
        "properties": {
          "code": {
            "type": "string",
            "description": "Stable machine-readable discriminator. Switch on this, not on `message`."
          },
          "message": {
            "type": "string"
          }
        }
      },
...
      "EvaluationStatus": {
        "type": "string",
        "description": "Where an evaluation got to.\n\nThe same three states a run has, for the same reason: a report that never\narrived and a report that failed are different facts.",
        "enum": [
          "running",
          "succeeded",
          "failed"
        ]
      },
...
    }
  },
...
}